feat: добавляет компонент MemoizedFullOffer
